requirements.txt
# Python dependencies for CursorTemplate project
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Web framework (uncomment if needed)
# flask>=2.3.0
# fastapi>=0.100.0
# uvicorn>=0.23.0

# Optional: Data processing (uncomment if needed)
# pandas>=2.0.0
# numpy>=1.24.0
# matplotlib>=3.7.0
